requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
Django==5.0.1
django-apscheduler==0.6.2
djangorestframework==3.14.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzlocal==5.2
